\name{census2000-package}
\alias{census2000-package}
\alias{census2000}
\docType{package}
\title{
Helper functions for the census2000-suite of packages
}
\description{
This package helps the user handle the spatial and demographic data available in \code{census2000tract}, \code{census2000cdp}, \code{census2000blkgrp}, and \code{census2000blk}
}
\details{
\tabular{ll}{
Package: \tab census2000\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2009-11-14\cr
License: \tab GPL Version 2 or newer\cr
LazyLoad: \tab yes\cr
}

}
\author{
Zack W. Almquist<almquist@uci.edu> \cr
Maintainer: Zack W. Almquist <almquist@uci.edu>
}
\note{
If you use this package and/or software manual in your work, a citation would be appreciated.  References to the current versions are:\cr

Zack W. Almquist (2009). census2000: US Census 2000 Suite of R Packages. R package version 0.01.

Please also cite the original data source \cr
Census 2000 Summary File 1 [name of state1 or United States]/prepared by the U.S. Census 
Bureau, 2001.
}

%\references{}
\keyword{uscensus2000}
%\seealso{}
%\examples{}
