\name{fips2000}
\alias{fips2000}
\docType{data}
\title{
FIPS for all levels in the year 2000
}
\description{
FIPS codes for all geographic levels of the US Census 2010. For use with API function.
}
\usage{data("fips2000")}
\source{
\url{https://www.census.gov/geo/reference/codes/cou.html}
}
\examples{
data(fips2000)
}
\keyword{datasets}
