\name{UScensus2000-package}
\alias{UScensus2000-package}
\alias{UScensus2000}
\docType{package}
\title{
Helper functions for the UScensus2000-suite of packages
}
\description{
This package helps the user handle the spatial and demographic data available in \code{UScensus2000tract}, \code{UScensus2000cdp}, \code{UScensus2000blkgrp}, and \code{UScensus2000blk}
}
\details{
\tabular{ll}{
Package: \tab UScensus2000\cr
Type: \tab Package\cr
Version: \tab 0.06\cr
Date: \tab 2009-11-14\cr
License: \tab GPL Version 2 or newer\cr
LazyLoad: \tab yes\cr
}

}
\author{
Zack W Almquist<almquist@uci.edu> \cr
Maintainer: Zack W Almquist <almquist@uci.edu>
}
\note{
If you use this package and/or software manual in your work, a citation would be appreciated.  References to the current versions are:\cr

Zack W Almquist (2009). UScensus2000: US Census 2000 Suite of R Packages. R package version 0.06.

Please also cite the original data source and the \code{\link[sp]{sp}} and \code{maptools}. \cr
Census 2000 Summary File 1 [name of state1 or United States]/prepared by the U.S. Census 
Bureau, 2001. 
\cr  \cr

Bivand RS, Pebesma EJ, Gomez-Rubio V (2008). Applied Spatial Data Analysis with R. Springer, New York, NY.


}

%\references{}
\keyword{usUScensus2000}
%\seealso{}
%\examples{}
